Extract locked course chapter in dedicated component
